Create Maximize Happiness of selected children
You are given an integer array happiness of length n, where each element represents the happiness value of a child standing in a queue. You are also given a positive integer k, representing the number of children you must select.

There are n children standing in a queue, and the iᵗʰ child has a happiness value of happiness[i].

Your goal is to select exactly k children in k turns while maximizing the total happiness collected.